Add files via upload
Project contains charts showing rolling weekly average for number of new cases and new deaths from Covid-19 for top (in terms of average) five countries.
Two types of charts are used to depict the trend among the countries. Line chart shows the trend for a given period of time and Map shows gradient wise intensity of cases and deaths amongst the filtered countries.